+++ cmd/gossamer/main_test.go
@@ -26,7 +26,6 @@ import (
 	"os"
 	"os/exec"
 	"strconv"
-	"strings"
 	"sync"
 	"syscall"
 	"testing"
@@ -34,9 +33,7 @@ import (
 	"time"
 
 	"github.com/ChainSafe/gossamer/dot"
-	"github.com/ChainSafe/gossamer/dot/state"
 	"github.com/ChainSafe/gossamer/lib/utils"
-	"github.com/dgraph-io/badger/v2"
 	"github.com/docker/docker/pkg/reexec"
 	"github.com/stretchr/testify/require"
 )
@@ -352,93 +349,3 @@ func TestBuildSpecCommandWithOutput(t *testing.T) {
 // TODO: TestInitCommand test "gossamer init" does not error
 
 // TODO: TestAccountCommand test "gossamer account" does not error
-
-func iterateDB(db *badger.DB, cb func(*badger.Item)) {
-	txn := db.NewTransaction(false)
-	itr := txn.NewIterator(badger.DefaultIteratorOptions)
-
-	for itr.Rewind(); itr.Valid(); itr.Next() {
-		cb(itr.Item())
-	}
-}
-
-func TestPruneState(t *testing.T) {
-	const (
-		bloomSize      = 256
-		retainBlockNum = 5
-	)
-
-	inDBPath := "../../tests/data/db"
-
-	pruner, err := newPruner(inDBPath, bloomSize, retainBlockNum)
-	require.NoError(t, err)
-
-	err = pruner.setBloomFilter()
-	require.NoError(t, err)
-
-	nonStorageKeys := make(map[string]interface{})
-	storageKeys := make(map[string]interface{})
-	itr := pruner.inputDB.NewIterator()
-	for itr.Next() {
-		key := string(itr.Key())
-		if !strings.HasPrefix(key, state.StoragePrefix) {
-			nonStorageKeys[key] = nil
-			continue
-		}
-		storageKeys[key] = nil
-	}
-	t.Log("Total keys in input DB", len(storageKeys)+len(nonStorageKeys), "storage keys", len(storageKeys))
-
-	// close the input DB because `prune-state` cmd will open it.
-	err = pruner.inputDB.Close()
-	require.NoError(t, err)
-
-	prunedDBPath := fmt.Sprintf("%s/%s", t.TempDir(), "badger")
-	t.Log("pruned DB path", prunedDBPath)
-
-	tt := runTestGossamer(t,
-		"prune-state",
-		"--basepath", inDBPath,
-		"--pruned-db-path", prunedDBPath,
-		"--bloom-size", "256",
-		"--retain-block", "5")
-
-	err = tt.cmd.Wait()
-	require.NoError(t, err)
-
-	prunedDB, err := badger.Open(badger.DefaultOptions(prunedDBPath))
-	require.NoError(t, err)
-
-	storageKeysPruned := make(map[string]interface{})
-	nonStorageKeysPruned := make(map[string]interface{})
-	getKeysPrunedDB := func(item *badger.Item) {
-		key := string(item.Key())
-		if strings.HasPrefix(key, state.StoragePrefix) {
-			key = strings.TrimPrefix(key, state.StoragePrefix)
-			storageKeysPruned[key] = nil
-			return
-		}
-		nonStorageKeysPruned[key] = nil
-	}
-	iterateDB(prunedDB, getKeysPrunedDB)
-	t.Log("Total keys in pruned DB", len(storageKeysPruned)+len(nonStorageKeysPruned), "storage keys", len(storageKeysPruned))
-	require.Equal(t, len(nonStorageKeysPruned), len(nonStorageKeys))
-
-	// Check all non storage keys are present.
-	for k := range nonStorageKeys {
-		_, ok := nonStorageKeysPruned[k]
-		require.True(t, ok)
-	}
-
-	// Check required storage keys exist.
-	for k := range storageKeys {
-		ok := pruner.bloom.contain([]byte(k))
-		if ok {
-			_, found := storageKeysPruned[k]
-			require.True(t, found)
-		}
-	}
-
-	err = prunedDB.Close()
-	require.NoError(t, err)
-}
